+++ tests/timeseries/test_ringbuffer.py
@@ -209,19 +209,19 @@ def test_gaps() -> None:  # pylint: disable=too-many-statements
     buffer = OrderedRingBuffer([0.0] * 5, ONE_SECOND)
     assert buffer.oldest_timestamp is None
     assert buffer.newest_timestamp is None
-    assert len(buffer) == 0
+    assert buffer.count_valid() == 0
     assert len(buffer.gaps) == 0
 
     buffer.update(Sample(dt(0), Quantity(0)))
     assert buffer.oldest_timestamp == dt(0)
     assert buffer.newest_timestamp == dt(0)
-    assert len(buffer) == 1
+    assert buffer.count_valid() == 1
     assert len(buffer.gaps) == 1
 
     buffer.update(Sample(dt(6), Quantity(0)))
     assert buffer.oldest_timestamp == dt(6)
     assert buffer.newest_timestamp == dt(6)
-    assert len(buffer) == 1
+    assert buffer.count_valid() == 1
     assert len(buffer.gaps) == 1
 
     buffer.update(Sample(dt(2), Quantity(2)))
@@ -229,49 +229,49 @@ def test_gaps() -> None:  # pylint: disable=too-many-statements
     buffer.update(Sample(dt(4), Quantity(4)))
     assert buffer.oldest_timestamp == dt(2)
     assert buffer.newest_timestamp == dt(6)
-    assert len(buffer) == 4
+    assert buffer.count_valid() == 4
     assert len(buffer.gaps) == 1
 
     buffer.update(Sample(dt(3), None))
     assert buffer.oldest_timestamp == dt(2)
     assert buffer.newest_timestamp == dt(6)
-    assert len(buffer) == 3
+    assert buffer.count_valid() == 3
     assert len(buffer.gaps) == 2
 
     buffer.update(Sample(dt(3), Quantity(np.nan)))
     assert buffer.oldest_timestamp == dt(2)
     assert buffer.newest_timestamp == dt(6)
-    assert len(buffer) == 3
+    assert buffer.count_valid() == 3
     assert len(buffer.gaps) == 2
 
     buffer.update(Sample(dt(2), Quantity(np.nan)))
     assert buffer.oldest_timestamp == dt(4)
     assert buffer.newest_timestamp == dt(6)
-    assert len(buffer) == 2
+    assert buffer.count_valid() == 2
     assert len(buffer.gaps) == 2
 
     buffer.update(Sample(dt(3), Quantity(3)))
     assert buffer.oldest_timestamp == dt(3)
     assert buffer.newest_timestamp == dt(6)
-    assert len(buffer) == 3
+    assert buffer.count_valid() == 3
     assert len(buffer.gaps) == 2
 
     buffer.update(Sample(dt(2), Quantity(2)))
     assert buffer.oldest_timestamp == dt(2)
     assert buffer.newest_timestamp == dt(6)
-    assert len(buffer) == 4
+    assert buffer.count_valid() == 4
     assert len(buffer.gaps) == 1
 
     buffer.update(Sample(dt(5), Quantity(5)))
     assert buffer.oldest_timestamp == dt(2)
     assert buffer.newest_timestamp == dt(6)
-    assert len(buffer) == 5
+    assert buffer.count_valid() == 5
     assert len(buffer.gaps) == 0
 
     buffer.update(Sample(dt(99), None))
     assert buffer.oldest_timestamp == dt(95)  # bug: should be None
     assert buffer.newest_timestamp == dt(99)  # bug: should be None
-    assert len(buffer) == 4  # bug: should be 0 (whole range gap)
+    assert buffer.count_valid() == 4  # bug: should be 0 (whole range gap)
     assert len(buffer.gaps) == 1
 
 
@@ -369,7 +369,7 @@ def test_len_ringbuffer_samples_fit_buffer_size() -> None:
             timestamp = start_ts + timedelta(seconds=index)
             buffer.update(Sample(timestamp, Quantity(float(sample_value))))
 
-        assert len(buffer) == len(test_samples)
+        assert buffer.count_valid() == len(test_samples)
 
 
 def test_len_with_gaps() -> None:
@@ -384,7 +384,7 @@ def test_len_with_gaps() -> None:
         buffer.update(
             Sample(datetime(2, 2, 2, 0, 0, i, tzinfo=timezone.utc), Quantity(float(i)))
         )
-        assert len(buffer) == i + 1
+        assert buffer.count_valid() == i + 1
 
 
 def test_len_ringbuffer_samples_overwrite_buffer() -> None:
@@ -412,7 +412,7 @@ def test_len_ringbuffer_samples_overwrite_buffer() -> None:
             timestamp = start_ts + timedelta(seconds=index)
             buffer.update(Sample(timestamp, Quantity(float(sample_value))))
 
-        assert len(buffer) == half_buffer_size
+        assert buffer.count_valid() == half_buffer_size
 
 
 def test_ringbuffer_empty_buffer() -> None: